Add resources + obj parser + tools + image manipulation + tests + GIF + BMP + JHelpImage + Animation

+++ Utilities/src/main/kotlin/khelp/utilities/math/Percent.kt
@@ -6,7 +6,7 @@ class Percent(percent : Double)
 {
     constructor(percent : Float) : this(percent.toDouble())
     constructor(percent : Int) : this(percent.toDouble() / 100.0)
-    constructor(percent : Long) : this(percent / 100.0)
+    constructor(percent : Long) : this(percent.toDouble() / 100.0)
 
     constructor(number : Int, total : Int) : this(number.toDouble() / total.toDouble())
     constructor(number : Long, total : Long) : this(number.toDouble() / total.toDouble())
